Add liquescimus
Despite this being the third song in this project from an anime, this is the first Japanese song. It's the first song with 3 languages (4 if you count Romaji as neither Japanese nor Latin). It's the first song with per-syllable syncing. It's the first song with vertical text.

For these reasons, this took longer to make than any other song so far.